+++ CYRKeyboardButton/CYRKeyboardButton.h
@@ -197,10 +197,16 @@ extern NSString *const CYRKeyboardButtonKeyPressedKey;
 @property (nonatomic, strong, nullable) NSArray *inputOptions;
 
 /**
+ An object that adopts the UIKeyInput protocol. When a key is pressed the key value is automatically inserted via the textInput object.
+ @discussion If the textInput object is not the first responder no text will be inserted.
+ */
+@property (nonatomic, weak, nullable) id<UIKeyInput> keyInput;
+
+/**
  An object that adopts the UITextInput protocol. When a key is pressed the key value is automatically inserted via the textInput object.
  @discussion If the textInput object is not the first responder no text will be inserted.
  */
-@property (nonatomic, weak, nullable) id<UITextInput> textInput;
+@property (nonatomic, weak, nullable) id<UITextInput> textInput __deprecated_msg("use keyInput");
 
 NS_ASSUME_NONNULL_END