+++ CYRKeyboardButton/CYRKeyboardButton.m
@@ -298,13 +298,28 @@ NSString *const CYRKeyboardButtonKeyPressedKey = @"CYRKeyboardButtonKeyPressedKe
     }
 }
 
+- (void)setKeyInput:(id<UIKeyInput>)keyInput
+{
+    NSAssert([keyInput conformsToProtocol:@protocol(UIKeyInput)], @"<CYRKeyboardButton> The key input object must conform to the UIKeyInput protocol!");
+    
+    [self willChangeValueForKey:NSStringFromSelector(@selector(keyInput))];
+    _keyInput = keyInput;
+    [self didChangeValueForKey:NSStringFromSelector(@selector(keyInput))];
+}
+
+@dynamic textInput;
+
 - (void)setTextInput:(id<UITextInput>)textInput
 {
     NSAssert([textInput conformsToProtocol:@protocol(UITextInput)], @"<CYRKeyboardButton> The text input object must conform to the UITextInput protocol!");
-    
+
     [self willChangeValueForKey:NSStringFromSelector(@selector(textInput))];
-    _textInput = textInput;
+    self.keyInput = textInput;
     [self didChangeValueForKey:NSStringFromSelector(@selector(textInput))];
+}
+
+- (id<UITextInput>)textInput {
+    return [_keyInput conformsToProtocol:@protocol(UITextInput)] ? (id<UITextInput>)_keyInput : nil;
 }
 
 #pragma mark - Internal - UI
@@ -398,17 +413,17 @@ NSString *const CYRKeyboardButtonKeyPressedKey = @"CYRKeyboardButtonKeyPressedKe
 {
     BOOL shouldInsertText = YES;
     
-    if ([self.textInput isKindOfClass:[UITextView class]]) {
+    if ([self.keyInput isKindOfClass:[UITextView class]]) {
         // Call UITextViewDelegate methods if necessary
-        UITextView *textView = (UITextView *)self.textInput;
+        UITextView *textView = (UITextView *)self.keyInput;
         NSRange selectedRange = textView.selectedRange;
         
         if ([textView.delegate respondsToSelector:@selector(textView:shouldChangeTextInRange:replacementText:)]) {
             shouldInsertText = [textView.delegate textView:textView shouldChangeTextInRange:selectedRange replacementText:text];
         }
-    } else if ([self.textInput isKindOfClass:[UITextField class]]) {
+    } else if ([self.keyInput isKindOfClass:[UITextField class]]) {
         // Call UITextFieldDelgate methods if necessary
-        UITextField *textField = (UITextField *)self.textInput;
+        UITextField *textField = (UITextField *)self.keyInput;
         NSRange selectedRange = [self textInputSelectedRange];
         
         if ([textField.delegate respondsToSelector:@selector(textField:shouldChangeCharactersInRange:replacementString:)]) {
@@ -417,23 +432,26 @@ NSString *const CYRKeyboardButtonKeyPressedKey = @"CYRKeyboardButtonKeyPressedKe
     }
     
     if (shouldInsertText == YES) {
-        [self.textInput insertText:text];
+        [self.keyInput insertText:text];
         
         [[NSNotificationCenter defaultCenter] postNotificationName:CYRKeyboardButtonPressedNotification object:self
                                                           userInfo:@{CYRKeyboardButtonKeyPressedKey : text}];
     }
 }
 
+// Note: caller must confirm that keyInput conforms to UITextInput
 - (NSRange)textInputSelectedRange
 {
-    UITextPosition *beginning = self.textInput.beginningOfDocument;
+    id<UITextInput> textInput = (id<UITextInput>)self.keyInput;
+
+    UITextPosition *beginning = textInput.beginningOfDocument;
     
-	UITextRange *selectedRange = self.textInput.selectedTextRange;
+	UITextRange *selectedRange = textInput.selectedTextRange;
 	UITextPosition *selectionStart = selectedRange.start;
 	UITextPosition *selectionEnd = selectedRange.end;
     
-	const NSInteger location = [self.textInput offsetFromPosition:beginning toPosition:selectionStart];
-	const NSInteger length = [self.textInput offsetFromPosition:selectionStart toPosition:selectionEnd];
+	const NSInteger location = [textInput offsetFromPosition:beginning toPosition:selectionStart];
+	const NSInteger length = [textInput offsetFromPosition:selectionStart toPosition:selectionEnd];
     
 	return NSMakeRange(location, length);
 }